Refactored folders to use  imports using barrel exports
