+++ src/components/cart/index.js
@@ -1,0 +1,5 @@
+export { default as Cart } from "./Cart";
+export { default as CartSidebar } from "./CartSidebar";
+export { default as CartItem } from "./CartItem";
+export { default as CartIcon } from "./CartIcon";
+export { default as QuantityCounter } from "./Quantity";
