+++ src/components/layout/index.js
@@ -1,0 +1,4 @@
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { default as Layout } from "./Layout";
+export { default as NavBar } from "./NavBar";
