+++ src/components/product/index.js
@@ -1,0 +1,2 @@
+export { default as Product } from "./Product";
+export { default as Reviews } from "./Reviews";
