+++ src/utils/calculateTotal.js
@@ -1,4 +1,3 @@
-// Calculate total price of items
 export const calculateTotal = (cart) => {
   return cart.reduce(
     (total, item) => total + item.discountedPrice * item.quantity,
